scm-core/src/test/java/sonia/scm/util/UrlBuilderTest.java
/**
 * Copyright (c) 2010, Sebastian Sdorra
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are met:
 *
 * 1. Redistributions of source code must retain the above copyright notice,
 *    this list of conditions and the following disclaimer.
 * 2. Redistributions in binary form must reproduce the above copyright notice,
 *    this list of conditions and the following disclaimer in the documentation
 *    and/or other materials provided with the distribution.
 * 3. Neither the name of SCM-Manager; nor the names of its
 *    contributors may be used to endorse or promote products derived from this
 *    software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
 * DISCLAIMED.  IN NO EVENT SHALL THE REGENTS OR CONTRIBUTORS BE LIABLE FOR ANY
 * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
 * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
 * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
 * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *
 * http://bitbucket.org/sdorra/scm-manager
 *
 */



package sonia.scm.util;

//~--- non-JDK imports --------------------------------------------------------

import org.junit.Test;

import static org.junit.Assert.*;

/**
 *
 * @author Sebastian Sdorra
 */
public class UrlBuilderTest
{

  /**
   * Method description
   *
   */
  @Test
  public void testAppendParameter()
  {
    UrlBuilder builder = new UrlBuilder("http://www.short.de");

    builder.appendParameter("i", 123).appendParameter("s", "abc");
    builder.appendParameter("b", true).appendParameter("l", 321l);
    assertEquals("http://www.short.de?i=123&s=abc&b=true&l=321",
                 builder.toString());
  }

  /**
   * Method description
   *
   */
  @Test
  public void testAppendUrlPart()
  {
    UrlBuilder builder = new UrlBuilder("http://www.short.de");

    builder.appendUrlPart("test");
    assertEquals("http://www.short.de/test", builder.toString());
  }

  /**
   * Method description
   *
   */
  @Test(expected = IllegalStateException.class)
  public void testState()
  {
    UrlBuilder builder = new UrlBuilder("http://www.short.de");

    builder.appendParameter("test", "123").appendUrlPart("test");
  }
}
